+++ CHANGELOG.md
@@ -3,6 +3,7 @@ This project attempts to conform to [SemVer](https://semver.org/) guidelines whe
 
 ## [v1.0.0]
 * [Adding discord build notifications](https://github.com/Ubunfu/mc-xp-bank/pull/10)
+* [Correcting withdraw API error status. Fixes #9](https://github.com/Ubunfu/mc-xp-bank/pull/11)
 
 ## [v0.2.1]
 * [Adding type checks on amounts passed into handlers](https://github.com/Ubunfu/mc-xp-bank/pull/7)